requirements.txt
# Chatterbox ES-LATAM Requirements
# For full setup (server + training)

# Core ML
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.26.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# TTS Model
chatterbox-tts

# Training
datasets>=2.21.0
tqdm
matplotlib>=3.5.0
huggingface_hub

# Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic-settings>=2.0.0
python-multipart  # For file uploads

# ONNX (for export/inference)
onnx>=1.14.0
onnxruntime>=1.16.0
